fix: router push urls
