+++ dashboard/pages/dashboard.tsx
@@ -1,0 +1,5 @@
+function Dashboard() {
+  return <div>Dashboard</div>;
+}
+
+export default Dashboard;
